Prawie wsyztsko na pierwszą wersję

+++ SyntezatorKrawczyka/src/main/java/com/jaebe/syntezatorkrawczyka/IodDo.java
@@ -3,9 +3,9 @@ package com.jaebe.syntezatorkrawczyka;
 /**
  * Created by Mateusz on 08.08.14.
  */
-interface IodDo
-{
-    public long getDelay();
-    public long getDlugosc();
+interface IodDo {
+    public int getDelay();
+
+    public int getDlugosc();
 
 }
